+++ setup.sh
@@ -17,14 +17,20 @@ sudo raspi-config nonint do_i2c 0
 sudo sed -i -e '/^dtparam=i2c_arm_baudrate/d' /boot/config.txt
 sudo sed -i -e 's/dtparam=i2c_arm=on$/dtparam=i2c_arm=on\ndtparam=i2c_arm_baudrate=100000/' /boot/config.txt
 
+# Enable the SPI support
+sudo raspi-config nonint do_spi 1
+sudo raspi-config nonint do_spi 0
+
 echo "Installing Raspberry Pi kernel headers"
 sudo apt-get install -y raspberrypi-kernel-headers
 
 echo "Installing the Python3 packages and related libs"
-sudo apt-get install python3-matplotlib
-sudo apt-get install python3-numpy
-sudo apt-get install libatlas-base-dev
+sudo apt-get install -y python3-matplotlib
+sudo apt-get install -y python3-numpy
+sudo apt-get install -y libatlas-base-dev
 
+echo  "Installing necessary packages for dev kit"
+sudo apt-get install -y libusb-1.0-0-dev libreadline-dev libncurses-dev
 
 # Build loader and insert it into the kernel
 if [ $# -ge 1 ] && [ $1 = "xvf3510" ] ; then
@@ -115,6 +121,6 @@ if [ $# -ge 1 ] && [ $1 = "xvf3510" ] ; then
 fi
 crontab $RPI_SETUP_DIR/resources/crontab
 
-echo "To enable I2S and I2C, this Raspberry Pi must be rebooted."
+echo "To enable I2S, I2C and SPI, this Raspberry Pi must be rebooted."
 
 popd > /dev/null