+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # VocalFusion Raspberry Pi Setup Change Log
 
+## 2.1.0
+
+  * Enable SPI interface
+  * Add installation of packages for dev kit
+
 ## 2.0.0
 
   * Use default I2C hardware